refactor: centralize user notion data retrieval
