+++ src/use-cases/books/add-book-to-notion.ts
@@ -1,15 +1,21 @@
 import { GoogleBooksService } from "../../lib/google-books";
 import { BookInfo } from "../../lib/google-books/types";
-import { PrismaUsersRepository } from "../../repositories/prisma/prisma-user";
 import { GetBookInfoUseCase } from "../books/get-info";
 import { AddBookItemUseCase } from "../notion/add-book-list-item-to-notion";
-import { GetUserDataUseCase } from "../user/get-user-data";
+import { UserService } from "../../services/user-service";
 
 export class AddBookToNotionUseCase {
+  constructor(
+    private userService: UserService = new UserService(),
+    private booksService: GoogleBooksService = new GoogleBooksService(
+      "AIzaSyCSD5DgEDcDEKA1zIWNw_a8VD8ag-ZUj6s"
+    )
+  ) {}
+
   async execute({ title, userId }: { title: string; userId: string }) {
     const { details } = await this.getBookInfos(title);
 
-    const userNotionData = await this.getUserNotionData(userId);
+    const userNotionData = await this.userService.getUserNotionData(userId);
 
     return await this.addToBooklist({
       bookInfos: details,
@@ -18,17 +24,8 @@ export class AddBookToNotionUseCase {
   }
 
   private async getBookInfos(title: string) {
-    const bookApiService = new GoogleBooksService(
-      "AIzaSyCSD5DgEDcDEKA1zIWNw_a8VD8ag-ZUj6s"
-    );
-    const getBookDetailsUseCase = new GetBookInfoUseCase(bookApiService);
+    const getBookDetailsUseCase = new GetBookInfoUseCase(this.booksService);
     return await getBookDetailsUseCase.execute({ title });
-  }
-
-  private async getUserNotionData(userId: string) {
-    const usersRepository = new PrismaUsersRepository();
-    const getUserNotionData = new GetUserDataUseCase(usersRepository);
-    return await getUserNotionData.execute(userId);
   }
 
   private async addToBooklist({
