+++ src/use-cases/tasks/add-task-to-notion.ts
@@ -1,21 +1,16 @@
-import { GetUserDataUseCase } from "../user/get-user-data";
-import { PrismaUsersRepository } from "../../repositories/prisma/prisma-user";
 import { AddTaskItemUseCase } from "../notion/add-task-item-to-notion";
+import { UserService } from "../../services/user-service";
 
 export class AddTaskToNotionUseCase {
+  constructor(private userService: UserService = new UserService()) {}
+
   async execute({ content, userId }: { content: string; userId: string }) {
-    const userNotionData = await this.getUserNotionData(userId);
+    const userNotionData = await this.userService.getUserNotionData(userId);
 
     return this.addToTasks({
       userNotionData,
       content,
     });
-  }
-
-  private async getUserNotionData(userId: string) {
-    const usersRepository = new PrismaUsersRepository();
-    const getUserNotionData = new GetUserDataUseCase(usersRepository);
-    return await getUserNotionData.execute(userId);
   }
 
   private async addToTasks({
